# server/docs/swagger.yaml
openapi: 3.0.0
info:
  title: VetCare Clinic API
  version: 1.0.0
  description: Veteriner Kliniği Yönetim Paneli Tüm API Dokümantasyonu
  contact:
    name: Deniz Can Özder
    email: contact@denizozder.com
servers:
  - url: http://localhost:5000/api
    description: Local Server
  - url: https://vetcare-api.onrender.com/api
    description: Production Server (Render)
...
components:
  schemas:
...
    Doctor:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        image:
          type: string
        imageFit:
          type: string

    # --- HİZMET ŞEMASI ---
...
paths:
...
  /doctors:
    get:
      summary: Tüm doktorları listele
      tags: [Doctors]
      responses:
        200:
          description: Başarılı
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
    post:
      summary: Yeni doktor ekle
      tags: [Doctors]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                title:
                  type: string
                summary:
                  type: string
                content:
                  type: string
                imageFit:
                  type: string
                image:
                  type: string
                  format: binary
      responses:
        201:
          description: Oluşturuldu
  /doctors/{id}:
    get:
      summary: Tek doktor getir
      tags: [Doctors]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        200:
          description: Detay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
    put:
      summary: Doktor güncelle
      tags: [Doctors]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                title:
                  type: string
                summary:
                  type: string
                content:
                  type: string
                imageFit:
                  type: string
                image:
                  type: string
                  format: binary
      responses:
        200:
          description: Güncellendi
    delete:
      summary: Doktor sil
      tags: [Doctors]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        200:
          description: Silindi

  # ==============================
  # HİZMET ROTALARI
  # ==============================
...